requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asn1crypto==0.24.0        # via cryptography
certifi==2019.9.11        # via requests
cffi==1.12.3              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.7         # via oauthlib
defusedxml==0.6.0         # via jira
hatchbuck==1.0.20
idna==2.8                 # via requests
jira==2.0.0
oauthlib[signedtoken]==3.1.0  # via jira, requests-oauthlib
pbr==5.4.3                # via jira
pycountry==19.8.18        # via hatchbuck
pycparser==2.19           # via cffi
pyjwt==1.7.1              # via oauthlib
python-dotenv==0.10.3
requests-oauthlib==1.2.0  # via jira
requests-toolbelt==0.9.1  # via jira
requests==2.22.0          # via hatchbuck, jira, requests-oauthlib, requests-toolbelt
six==1.12.0               # via cryptography, jira
urllib3==1.25.6           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via jira
